refactor(dette): J3 lot B — unconvert, goconst et le nommage, sous gate des artefacts
Trois linters mecaniques, traites un par un, chacun avec son gate complet.

unconvert (3) — verifie sur pieces qu aucune largeur de bits n est en jeu :
velocityMagBits est deja declare uint, ReadBits rend uint64, b2i rend int. Les
trois conversions etaient des no-op de type, pas des troncatures.

goconst (3) — deux etiquettes de composant citees a plus de deux endroits passent
en constantes (compObjectBodyVitality, compWeaponStateTypeInfo). Les autres restent
en litteral : les centraliser toutes creerait une table d indirection sans lecteur.
Pas de test de grep ajoute, et le commentaire le dit : le risque que porte une
etiquette dupliquee — deux branches qui ne consomment pas le meme nombre de bits —
est deja couvert par TestCaptureConsumesSameBitsAsDispatch.

revive var-naming (3) — consumeShotId -> consumeShotID, consumeId2 -> consumeID2,
consumeForgePlayerDataEditedObjectsIds -> ...IDs. Les 3 argument-limit restants
sont HORS perimetre de la passe mecanique (ils demandent de regrouper des
parametres, pas de renommer) — consignes pour la session 2.

GATE apres chaque linter : les 3 artefacts de rejeu reconstruits, tous les chiffres
du §5 de PLAN_RECONCILIATION_BRANCHES.md identiques a l unite pres (99/29221,
475/519, 90/105, 70/70, 439 projectiles, 184 etats d inventaire sur 24 images-cles,
10223 emprises ; 1862/2154 ; 2312/2879), plus replay + filmdec + killsource verts.

+++ apps/go-api/internal/analysis/filmdec/components_cubemap.go
@@ -68,7 +68,7 @@ func DecodeVelocityMagnitude(scale uint64, width uint) float32 {
 // as FUN_14076d528 does (unit dir * FUN_14076d6dc magnitude). This is the datum that
 // dead-reckons a player's position through keep-baseline frames.
 func DecodeVelocity(packedDir, scale uint64) [3]float32 {
-	d := DecodeDynPrecDir(packedDir, uint(velocityMagBits))
+	d := DecodeDynPrecDir(packedDir, velocityMagBits)
 	m := DecodeVelocityMagnitude(scale, velScaleBitsDef)
 	return [3]float32{d[0] * m, d[1] * m, d[2] * m}
 }
